split azureSpeech and webSpeech functions into separate files
